Carrito

+++ producto/models.py
@@ -51,7 +51,8 @@ class Producto(models.Model):
     )
 
     descripcion = models.TextField()
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.FloatField()
+    #precio = models.DecimalField(max_digits=10, decimal_places=2)
     imagen = models.URLField()
     stock = models.IntegerField()
